Hexagonal Architecture Sample - Loan Application and Approval System

+++ src/main/java/com/dominikcebula/samples/loans/application/port/in/dto/ApplicantDTO.java
@@ -2,6 +2,6 @@ package com.dominikcebula.samples.loans.application.port.in.dto;
 
 import java.time.LocalDate;
 
-public record ApplicantDTO(int id, String firstName, String lastName, LocalDate birthDate, int creditScore,
+public record ApplicantDTO(Long id, String firstName, String lastName, LocalDate birthDate, int creditScore,
                            EmploymentDTO employment, String email, String phoneNumber) {
 }
