+++ src/main/java/com/dominikcebula/samples/loans/application/port/in/dto/EmploymentDTO.java
@@ -1,5 +1,5 @@
 package com.dominikcebula.samples.loans.application.port.in.dto;
 
-public record EmploymentDTO(int id, String name, String industry, MoneyDTO yearlyIncome, int numberOfEmployees,
+public record EmploymentDTO(Long id, String name, String industry, MoneyDTO yearlyIncome, int numberOfEmployees,
                             String email, String phoneNumber, String website) {
 }
